Move the Planck 2018 CLI tests out of the app gate
* Split test_generate_planck and test_generate_planck_test into their own file
* Place it outside tests/python/numcosmo_py/app, whose name alone gates on --run-app
