+++ tests/python/numcosmo_py/app/test_numcosmo_app.py
@@ -41,7 +41,6 @@ from numcosmo_py import Ncm, Nc
 from numcosmo_py.app.loading import load_catalog
 from numcosmo_py.app import app
 from numcosmo_py.app.esmcmc import IniSampler, Parallelization
-from numcosmo_py.app.generate import Planck18Types
 from numcosmo_py.interpolation.stats_dist import CrossValidationMethod
 from numcosmo_py.sampling import FitRunner, FitGradType, FitRunMessages, FisherType
 from numcosmo_py.interpolation.stats_dist import (
@@ -124,12 +123,6 @@ def fixture_interpolation_kernel(request) -> str:
 )
 def fixture_calibration_method(request) -> str:
     """Return calibration method."""
-    return request.param
-
-
-@pytest.fixture(name="planck18_type", params=[e.value for e in Planck18Types])
-def fixture_planck18_type(request) -> str:
-    """Return planck18 type."""
     return request.param
 
 
@@ -1373,34 +1366,3 @@ def test_run_mcmc_apes_get_best_fit_requires_output(simple_experiment):
     result = runner.invoke(app, ["catalog", "get-best-fit", catalog.as_posix()])
     assert result.exit_code != 0
     assert "Output file not defined" in result.output
-
-
-def test_generate_planck(tmp_path, planck18_type):
-    """Test run theory vector."""
-    tmp_file = tmp_path / "planck_generated1.yaml"
-
-    result = runner.invoke(
-        app,
-        ["generate", "planck18", tmp_file.as_posix(), "--data-type", planck18_type],
-    )
-
-    if result.exit_code != 0:
-        raise result.exception
-
-
-def test_generate_planck_test(tmp_path, planck18_type):
-    """Test run theory vector."""
-    tmp_file = tmp_path / "planck_generated2.yaml"
-
-    result = runner.invoke(
-        app,
-        ["generate", "planck18", tmp_file.as_posix(), "--data-type", planck18_type],
-    )
-
-    if result.exit_code != 0:
-        raise result.exception
-
-    result = runner.invoke(app, ["run", "test", tmp_file.as_posix()])
-
-    if result.exit_code != 0:
-        raise result.exception
